requirements.txt
fastapi
uvicorn
python-dotenv
websockets
jinja2
requests
